moved emane dependent tests into subdir, moved pytest config into setup.cfg, added CoreError and made use of in session.py
