+++ daemon/core/__init__.py
@@ -31,3 +31,11 @@ class CoreCommandError(subprocess.CalledProcessError):
 
     def __str__(self):
         return "Command(%s), Status(%s):\n%s" % (self.cmd, self.returncode, self.output)
+
+
+class CoreError(Exception):
+    """
+    Used for errors when dealing with CoreEmu and Sessions.
+    """
+
+    pass
