+++ daemon/setup.cfg
@@ -15,3 +15,6 @@ max-line-length=100
 max-complexity=26
 select=B,C,E,F,W,T4
 exclude=*_pb2*.py,utm.py,doc,build
+
+[tool:pytest]
+norecursedirs=distributed emane
