--- daemon/tests/pytest.ini
@@ -1,2 +1,0 @@
-[pytest]
-norecursedirs = distributed
